requirements.txt
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
coverage==5.2.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-heroku==0.3.1
django-import-export==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==4.1.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
MarkupPy==1.14
more-itertools==8.4.0
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
